Added an EEPROM parser and EEPROM builder to handle configuration of the device (not fully tested for now)

+++ code/mixerlib/Src/maths/maths.c
@@ -26,7 +26,7 @@
  * DEFINES
  * -----------------------------------------------------------------
  */
-#define ADC_RESOLUTION (double)(ADC_MAX_VOLTAGE/ ((1 << ADC_RESOLUTION_BITS) - 1))
+#define ADC_RESOLUTION (float)(ADC_MAX_VOLTAGE/ ((1 << ADC_RESOLUTION_BITS) - 1))
 
 /* -----------------------------------------------------------------
  * FUNCTIONS
@@ -36,18 +36,18 @@
 /**
  * ADC Conversions
  */
-inline int ADC2Double(uint16_t code, double *val)
+inline int ADC2Double(uint16_t code, float *val)
 {
 	if (code > (1 << ADC_RESOLUTION_BITS))
 	{
 		return -1;
 	}
 
-	*val = (double)code * ADC_RESOLUTION;
+	*val = (float)code * ADC_RESOLUTION;
 	return 0;
 }
 
-inline int Double2ADC(double val, uint16_t *code)
+inline int Double2ADC(float val, uint16_t *code)
 {
 	if ((0 > val) | (val > ADC_MAX_VOLTAGE))
 	{
@@ -61,13 +61,13 @@ inline int Double2ADC(double val, uint16_t *code)
 /**
  * S1.6 conversion
  */
-inline int S162Double(int8_t code, double *val)
+inline int S162Double(int8_t code, float *val)
 {
 	*val = code / 64.0;
 	return 0;
 }
 
-inline int Double2S16(double val, int8_t *code)
+inline int Double2S16(float val, int8_t *code)
 {
 	if ((-1.5 > val) | (val > 1.5))
 	{
